backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.4.2
tensorflow==2.14.0
pillow==10.1.0
numpy==1.24.3
openai==1.3.5
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
cors==1.0.1 